[refactoring][storage] Rename SSD Interface file
